split examples to one per file, jwx naming convention
