+++ examples/content_example_test.go
