+++ examples/metaschema_test.go
@@ -3,7 +3,6 @@ package examples_test
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"testing"
 
 	"github.com/lestrrat-go/json-schema/meta"
@@ -142,54 +141,4 @@ func BenchmarkMetaSchemaValidation(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleValidator() {
-	ctx := context.Background()
-
-	// Get the pre-compiled meta-schema validator
-	validator := meta.Validator()
-
-	// Example: validate a simple string schema
-	stringSchema := map[string]any{
-		"type":      "string",
-		"minLength": 1,
-		"maxLength": 100,
-	}
-
-	_, err := validator.Validate(ctx, stringSchema)
-	if err != nil {
-		// Schema is not valid according to JSON Schema meta-schema
-		panic(err)
-	}
-
-	// Schema is valid!
-	fmt.Println("Schema is valid")
-	// Output:
-	// Schema is valid
-}
-
-func ExampleValidate() {
-	ctx := context.Background()
-
-	// Example: validate an object schema using the convenience function
-	objectSchema := map[string]any{
-		"type": "object",
-		"properties": map[string]any{
-			"name": map[string]any{"type": "string"},
-			"age":  map[string]any{"type": "integer", "minimum": 0},
-		},
-		"required": []string{"name"},
-	}
-
-	err := meta.Validate(ctx, objectSchema)
-	if err != nil {
-		// Schema is not valid according to JSON Schema meta-schema
-		panic(err)
-	}
-
-	// Schema is valid!
-	fmt.Println("Object schema is valid")
-	// Output:
-	// Object schema is valid
 }
