+++ examples/validator_builder_example_test.go
@@ -7,12 +7,12 @@ import (
 	"github.com/lestrrat-go/json-schema/validator"
 )
 
-// Example_validatorBuilder constructs a validator directly with the validator
+// Example_validator_builder constructs a validator directly with the validator
 // package's builders, instead of compiling a *schema.Schema. This is the
 // lower-level path: useful when the constraints are known in Go and there is no
 // schema document to load. (It has no JSON form — that is exactly what loading a
 // schema and calling validator.Compile is for.)
-func Example_validatorBuilder() {
+func Example_validator_builder() {
 	v := validator.String().
 		MinLength(3).
 		MaxLength(10).
